fix bug

+++ sao_paulo/src/main/java/org/eqasim/sao_paulo/mode_choice/utilities/estimators/SaoPauloPTUtilityEstimator.java
@@ -35,7 +35,7 @@ public class SaoPauloPTUtilityEstimator  extends PtUtilityEstimator{
 	}
 	
 	protected double estimateAgeUtility(Person person) {
-		return (double) person.getAttributes().getAttribute("age") <= 16 ? parameters.spPT.alpha_age : 0.0;
+		return (int) person.getAttributes().getAttribute("age") <= 16 ? parameters.spPT.alpha_age : 0.0;
 	}
 
 	@Override
